Create utterances.json
Hi there Jason @jodonn We are starting to look to onboard PostgresSQL  into Chatbot. In order for Chatbot to 'know' about the FAQs and troubleshooting questions we have created utterances to support this new troubleshooting question that Jeanmarie submitted. If you could please approve and merge this utterance file we would appreciate it. Once completed, if this file could be copied to publish, so that it can be pushed to production - we will then build the intents in Chatbot. Please reach out with any questions, Claire
There is a more detailed write up for our Chatbot here - https://test.cloud.ibm.com/docs/writing?topic=writing-utterances