Delete utterances.json
The Cloud Assistant returns this topic as an answer more than it should, for subject areas unrelated to connectivity, because it is the only Troubleshooting/FAQ for this service.  Until there are more Troubleshooting topics or FAQs on which to train the Cloud Assistant for distinct intents, please remove this utterances.json file.  Thank you.